backend/pyproject.toml
[project]
name = "gamma-backend"
version = "0.1.0"
description = "Gamma backend: FastAPI + PostgreSQL (schema-per-tenant) + Celery"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic[email]>=2.9.0",
    "pydantic-settings>=2.6.0",
    "sqlalchemy[asyncio]>=2.0.35",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "celery[redis]>=5.4.0",
    "redis>=5.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib>=1.7.4",
    "bcrypt<4.0",
    "python-multipart>=0.0.12",
    "httpx>=0.27.0",
    "jinja2>=3.1.4",
    "pyotp>=2.9.0",
    "webauthn>=2.2.0",
    "orjson>=3.10.0",
    "structlog>=24.4.0",
    "aiosmtplib>=3.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "hypothesis>=6.115.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "factory-boy>=3.3.0",
    "faker>=30.0.0",
    "aiosqlite>=0.20.0",
]
prod = [
    # Real vendor SDKs used by the M1 wrappers in production.
    # Dev defaults to stub implementations and does not need these.
    "google-cloud-aiplatform>=1.71.0",
    "google-cloud-storage>=2.18.0",
    "google-cloud-secret-manager>=2.20.0",
    "weasyprint>=63.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM", "RUF"]

[tool.ruff.lint.per-file-ignores]
"app/features/*/models.py" = ["N815"]
"tests/*" = ["B008"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra --strict-markers"
markers = [
    "financial: tests that cover financial math (require 100% coverage)",
    "integration: tests that need a real Postgres (skip without DATABASE_URL)",
]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]
ignore_missing_imports = true
